Just a hack that makes white noise video
This hack makes white noise video, but need about 5 minutes on a reasonable hardware to start up. 
No multithreading, which should speed things up, but that is for a day in the future.
It uses a buffer to create 180 random images with white and black pixels, which it copy to the screen 30 times per second when running.